src/django/requirements.txt
aiokafka==0.8.0
base32-crockford==0.3.0
boto3==1.24.23
coreapi==2.3.3
coverage
dedupe==1.9.4
debugpy==1.5.1
defusedxml==0.6.0
django-amazon-ses==4.0.1
django-cors-headers==3.13.0
django-ecsmanage==2.0.1
django-elasticache==1.0.3
django-extensions==3.1.4
django-jsonview==2.0.0
django-rest-auth[with_social]==0.9.5
django-simple-history==3.1.1
django-spa==0.3.5
django-storages==1.13.1
django-waffle==2.5.0
django-watchman==1.3.0
drf-yasg==1.20.0
djangorestframework-gis==1.0.0
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
mercantile==1.1.2
openpyxl==3.0.9
pycodestyle==2.8.0
pyflakes==2.4.0
pylibmc==1.6.0
pytz==2018.7
pymemcache==3.5.2
requests==2.28.1
rollbar==0.16.3
thefuzz==0.19.0
Unidecode==1.0.23
django-slowtests
django-allauth==0.54.0